requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
PyMySQL
python-dotenv
cryptography